# apps/backend/docs/openapi-internal.yaml
openapi: 3.1.0
info:
  title: Transcendence Internal API
  version: 0.1.0
  description: |
    Internal API used by the frontend. These endpoints are intentionally kept
    separate from the future public API specification.

    The backend process receives routes without the `/api` prefix, while the
    reverse proxy exposes the routes below under `/api`.
servers:
  - url: /
    description: Current host (the reverse proxy exposes the API below /api)
...
paths:
...
  /api/internal/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - betterAuthSession: []
      responses:
        '200':
          description: Users visible to authenticated clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/internal/admin/users:
    get:
      tags: [Admin users]
      summary: Search users as an administrator
      operationId: searchAdminUsers
      description: Requires an authenticated user with the `admin` role.
      security:
        - betterAuthSession: []
      parameters:
        - $ref: '#/components/parameters/AdminUserSearchQuery'
        - $ref: '#/components/parameters/AdminUserRoleQuery'
      responses:
        '200':
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AdminRequired'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/internal/admin/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Admin users]
      summary: Get a user as an administrator
      operationId: getAdminUser
      description: Requires an authenticated user with the `admin` role.
      security:
        - betterAuthSession: []
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AdminRequired'
        '404':
          description: "User not found (error.code: USER_NOT_FOUND)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                userNotFound:
                  $ref: '#/components/examples/UserNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Admin users]
      summary: Update a user's name
      operationId: updateAdminUser
      description: Requires an authenticated user with the `admin` role.
      security:
        - betterAuthSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchAdminUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AdminRequired'
        '404':
          description: "User not found (error.code: USER_NOT_FOUND)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                userNotFound:
                  $ref: '#/components/examples/UserNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Admin users]
      summary: Delete a user
      operationId: deleteAdminUser
      description: Requires an authenticated user with the `admin` role. An administrator cannot delete their own account.
      security:
        - betterAuthSession: []
      responses:
        '204':
          description: User deleted
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AdminRequired'
        '404':
          description: "User not found (error.code: USER_NOT_FOUND)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                userNotFound:
                  $ref: '#/components/examples/UserNotFound'
        '409':
          description: An administrator cannot delete their own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  examples:
    UserNotFound:
      summary: User not found
      value:
        error:
          code: USER_NOT_FOUND
          message: User not found
...
  securitySchemes:
    betterAuthSession:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie issued by Better Auth.
  parameters:
...
    AdminUserSearchQuery:
      name: q
      in: query
      description: Case-insensitive search in user name and email
      required: false
      schema:
        type: string
        maxLength: 200
    AdminUserRoleQuery:
      name: role
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/UserRole'
...
    UserId:
      name: userId
      in: path
      required: true
      description: Better Auth user identifier
      schema:
        type: string
  responses:
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    AdminRequired:
      description: Administrator privileges are required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable application error code
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request
        details:
          description: Additional error details. Validation errors include fieldErrors and formErrors.
          oneOf:
            - $ref: '#/components/schemas/ValidationDetails'
            - {}
    ValidationDetails:
      type: object
      properties:
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        formErrors:
          type: array
          items:
            type: string
    UserRole:
      type: string
      enum: [admin, user]
    UserImage:
      type: [string, 'null']
      description: Avatar URL or another image URL stored by Better Auth
    UserSummary:
      type: object
      required: [id, name, image]
      properties:
        id:
          type: string
        name:
          type: string
        image:
          $ref: '#/components/schemas/UserImage'
...
    AdminUser:
      type: object
      required:
        - id
        - name
        - email
        - emailVerified
        - image
        - role
        - banned
        - banReason
        - banExpires
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
        image:
          $ref: '#/components/schemas/UserImage'
        role:
          $ref: '#/components/schemas/UserRole'
        banned:
          type: boolean
        banReason:
          type: [string, 'null']
        banExpires:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PatchAdminUserRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...